Fix fast refresh compatibility for person profile
