+++ frontend/src/employee-frontend/components/child-information/ChildApplications.tsx
@@ -18,8 +18,8 @@ import { ChildContext } from '../../state'
 import { useTranslation } from '../../state/i18n'
 import { UIContext } from '../../state/ui'
 import { RequireRole } from '../../utils/roles'
+import { inferApplicationType } from '../application'
 import { renderResult } from '../async-rendering'
-import { inferApplicationType } from '../person-profile/PersonApplications'
 import { DateTd, NameTd, StatusTd } from '../person-profile/common'
 
 import CreateApplicationModal from './CreateApplicationModal'
